Update to Grails 3
