--- test/resources/certs/EquifaxSecureGlobaleBusinessCA-1.crt
@@ -1,20 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDLjCCApegAwIBAgIDDBLjMA0GCSqGSIb3DQEBBQUAME4xCzAJBgNVBAYTAlVT
-MRAwDgYDVQQKEwdFcXVpZmF4MS0wKwYDVQQLEyRFcXVpZmF4IFNlY3VyZSBDZXJ0
-aWZpY2F0ZSBBdXRob3JpdHkwHhcNMDkwNzIyMTcwMDI1WhcNMTAxMDIyMjEwNzM0
-WjCBuDELMAkGA1UEBhMCVVMxFzAVBgNVBAoUDiouZGV2LmphdmEubmV0MRMwEQYD
-VQQLEwpHVDg4MjQ1NjQ1MTEwLwYDVQQLEyhTZWUgd3d3LnJhcGlkc3NsLmNvbS9y
-ZXNvdXJjZXMvY3BzIChjKTA5MS8wLQYDVQQLEyZEb21haW4gQ29udHJvbCBWYWxp
-ZGF0ZWQgLSBSYXBpZFNTTChSKTEXMBUGA1UEAxQOKi5kZXYuamF2YS5uZXQwgZ8w
-DQYJKoZIhvcNAQEBBQADgY0AMIGJAoGBALVWD2FBlAPBrdRbgZ+EBp07+Z6hr046
-Ba9LJC6eykjRHXKEFR0finMFzmtmhfDoXPKpSB9pcnl/fMXy7UQX2kEbyPItGZfY
-t+YsGez59mUjU8XsdmA6dTaXb+vtP/R6YgMOrufVDRKEqPASYQ26lffakBHxgSJd
-z6Uei8kJ72ShAgMBAAGjga4wgaswDgYDVR0PAQH/BAQDAgTwMB0GA1UdDgQWBBRB
-hXi+DbU3DAN5vodRZyDjT6bp/DA6BgNVHR8EMzAxMC+gLaArhilodHRwOi8vY3Js
-Lmdlb3RydXN0LmNvbS9jcmxzL3NlY3VyZWNhLmNybDAfBgNVHSMEGDAWgBRI5mj5
-K9KylddH2CMgEE8zmJCf1DAdBgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIw
-DQYJKoZIhvcNAQEFBQADgYEAlTBnlfWDgVFdNsKzDI05N38oPaQh2O8JcFtI0vm+
-rhwJYj6vjGdeH1bspIJ+f6ZjciOiaADwermT0QyFTsW/6PRpi5rd4Z+iv2tR4ieB
-/2ECgo/GQPs39B40mVoXPcZNoDGKWBAhDbYC+ehPe9TmSoqL4J5wchvR0oxlj1L7
-kVU=
------END CERTIFICATE-----
